TA6508 Do not authenticate authomatically on every request
Issue: the implemented ServerAuthModule attempts to authenticate every not
authenticated request by the configured oauth provider. This behaviour does
not allow to authenticate reacting to a specific client action (request).

Solution: introduce a new configuration parameter: module.loginRequestPattern
which indicates that only requests matching this pattern
should be authenticated with the provider. If the paramter is not specified
the old behaviour would be applied. Also, another configuration
parameter "module.forwardToIfNotAuthenticated" is introduced to specify
the landing page for all requests which are not autenticated, and also
do not follow the pattern.

Note: the ServerAuthModule still does not support multiple providers since
all of the configured modules share the same configuration.

Implemented by mwo
Reviewed by pp